Bug Fix: Remove extra commas from embedded struct field
Upon encoding a struct embedded with an
uninitialized pointer of another struct extra commas
were being generated.

This was primarily because of not rolling back
the comma we add for each field in case of encode failure.
Rerranged the if conditional to mitigate the issue.

Added corresponding unit test as well.

For detailed description of issue refer -
https://github.com/segmentio/encoding/issues/41